KVM: PPC: Book3S HV: Add a fairer algorithm to grab host core in find_available_hostcore()
When one core try to take advantage of the guest->host IPI mechanism
to ask some available host core to do something like VCPU wakeup, it will
always grab the host core which is close to it currently.

This patch consider using a fairer (Round-Robin) algorithm to grab host
core in find_available_hostcore(). It could be helpful to distributes the
IPIs better.

Signed-off-by: Yongji Xie <xyjxie@linux.vnet.ibm.com>