KVM: PPC: Book3S HV: Send IPI to host core to handle fast mmio
This patch adds support to real-mode KVM to search for a core
running in the host partition and send it an IPI message with
fast mmio gpa to be handled. Then we wait for host core to
finish handling the mmio access in real mode. If the fast mmio
access is completed by host core, we could avoid the context
switch and return to guest directly. This would improve IO
performance of virtio devices, especially when SMT is on.

In current implenment, we apply this optimization to each mmio
write. This may slightly slow down some emulated devices.

Signed-off-by: Yongji Xie <xyjxie@linux.vnet.ibm.com>